+++ src/main/java/net/runelite/client/plugins/microbot/microbotdashboardplus/MicrobotDashboardPlusPlugin.java
@@ -64,7 +64,7 @@ import java.awt.image.BufferedImage;
 @Slf4j
 public class MicrobotDashboardPlusPlugin extends Plugin {
 
-    public static final String version = "0.3.0";
+    public static final String version = "0.3.1";
 
     @Inject
     private MicrobotDashboardPlusConfig config;